Added Not found page. (#92)
* Removed catch all route to allow 404 to fall through to error element.
Layout updated to accept optional `outlet` which is rendered in favour of default where supplied. This allows reuse of the existing layout in the errorElement for a consistent look and feel in error states.
The `NotFound` and `ComingSoon` components were essentially the same and so have been extracted in to a `ContentWrapper` container which renders children. This allows for greater reuse.
The error component also now contains logic to identify and render the error message