Add shell.nix providing a dev environment
This commit adds a shell.nix file which can be be used to provide a
development environment, if one has [nix](https://nixos.org/)
installed. (To enter the environment, run `nix-shell` at the repository
root.)

The definition of the development environment is provided by
[guardian-nix](https://github.com/guardian/guardian-nix/): given the
list of separate commands required to run the project, it provides a
command `run-all-tmux` which runs all of the commands in a new tmux
session.

This commit also uses [niv](https://github.com/nmattia/niv) to “pin”
nixpkgs to a specific commit: this ensures that the environment is the
same across different machines, rather than depending what version of
nixpkgs (the nix package repository) that machine is currently pointed
to.